requirements.txt
openai
langchain
langgraph
chromadb
tavily-python
python-dotenv
streamlit
Flask
agentlightning
requests
beautifulsoup4
pandas
python-docx
reportlab
PyMuPDF
google-api-python-client
google-auth
google-auth-oauthlib
spacy
dateparser
httpx
